Universal Table
Im working now at the thead to make it editable instantly without needing to open the form to edit it, after then I have to work on the tables. What I need to do there is -  Everytime I choose the universal template I need to open a new table, a blank one while the old one is saved in the dropdown.